+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- Rename internal ``IndexedSeries`` to just ``Series``
 - Repeat previous value when there is a gap in the input
 
 ## [0.4.0] - 2023-10-23